Implement Add Project Button
Implement button and logic for user to add projects. Reorganize
relevant code (which is most of it) to appLogic module. The button
creates a short form for user to input a title and description,
with a submit button to add it to the projects list. The sidebar
content is then refreshed to properly display everything.