Implement Create New Project Functionality
Each listed project is a button. Clicking each project button now
selects the project to display on the main section of the page with
the respective todos underneath. A custom data-index attribute is
included on each sidebar project div in order to select the proper
project in the projects array when using the buttons.